fix(cc): add compability on Windows
