+++ umjunsik-lang-cc/clean.cmd
@@ -1,0 +1,3 @@
+@echo off
+del /q ummc.exe
+del /q src\*.o
